Move Trace to common to mitigate changes
